@@ -73,7 +73,7 @@ open class NetworkRequest {
                                           headers: [NetworkHeader]? = nil,
                                           query: [NetworkQuery]? = nil,
                                           jsonBody: T? = nil,
-                                          encoder: JSONEncoder = JSONEncoder()) throws {
+                                          encoder: JSONEncoder = JSONEncoder(outputFormatting: .sortedKeys)) throws {
         var httpBodyData: Data?
         if let jsonBody = jsonBody {
             do {
@@ -98,9 +98,58 @@ open class NetworkRequest {
                                           headers: [NetworkHeader]? = nil,
                                           query: [NetworkQuery]? = nil,
                                           jsonBody: T? = nil,
-                                          encoder: JSONEncoder = JSONEncoder()) throws {
+                                          encoder: JSONEncoder = JSONEncoder(outputFormatting: .sortedKeys)) throws {
         guard let url = URL(string: url) else { throw NetworkError.request(.invalidURL) }
         try self.init(url: url, httpMethod: httpMethod, headers: headers, query: query, jsonBody: jsonBody)
+    }
+    
+    /// Crea un `NetworkRequest` a partir de unos parámetros. Este inicializador añade la cabecera `Content-Type: multipart/form-data` automáticamente a la petición
+    ///  - Parameters:
+    ///  - url: url de la petición
+    ///  - httpMethod: método http de la petición
+    ///  - headers: cabeceras de la petición
+    ///  - query: parámetros de tipo query de la petición
+    ///  - multipartForm: parámetros de tipo `NetworkMultipartFormDataConvertible` que serán añadidos al body de la petición
+    public convenience init(url: String,
+                            httpMethod: NetworkHttpMethod,
+                            headers: [NetworkHeader]? = nil,
+                            query: [NetworkQuery]? = nil,
+                            multipartForm: [NetworkMultipartFormDataConvertible]) throws {
+        guard let url = URL(string: url) else { throw NetworkError.request(.invalidURL) }
+        try self.init(url: url, httpMethod: httpMethod, headers: headers, query: query, multipartForm: multipartForm)
+    }
+    /// Crea un `NetworkRequest` a partir de unos parámetros. Este inicializador añade la cabecera `Content-Type: multipart/form-data` automáticamente a la petición
+    /// - Parameters:
+    ///  - url: url de la petición
+    ///  - httpMethod: método http de la petición
+    ///  - headers: cabeceras de la petición
+    ///  - query: parámetros de tipo query de la petición
+    ///  - multipartForm: parámetros de tipo `NetworkMultipartFormDataConvertible` que serán añadidos al body de la petición
+    public convenience init(url: URL,
+                            httpMethod: NetworkHttpMethod,
+                            headers: [NetworkHeader]? = nil,
+                            query: [NetworkQuery]? = nil,
+                            multipartForm: [NetworkMultipartFormDataConvertible]) throws {
+        var httpBodyData: Data = Data()
+        let boundary = "Boundary-\(UUID().uuidString)"
+        guard let boundaryData = "--\(boundary)\r\n".data(using: .utf8) else {
+            throw NetworkError.request(.stringEncodingError)
+        }
+        if !multipartForm.isEmpty {
+            httpBodyData.append(
+                try multipartForm.reduce(into: Data()) { (result, item) in
+                    do {
+                        let data = try item.data(boundary: boundary)
+                        result.append(data)
+                    } catch {
+                        throw NetworkError.request(.encodeError(error))
+                    }
+                }
+            )
+            httpBodyData.append(boundaryData)
+        }
+        
+        try self.init(url: url, httpMethod: httpMethod, headers: [NetworkHeader(key: "Content-Type", value: "multipart/form-data; boundary=\(boundary)")] + (headers ?? []), query: query, httpBody: httpBodyData)
     }
     
     /// Crea un `NetworkRequest` a partir de un `URLRequest`
